+++ tests/tests.rs
@@ -742,7 +742,7 @@ fn test_unsized_generic_params() {
 }
 
 #[test]
-fn test_field_setter_transform() {
+fn test_field_setter_transform_closure() {
     #[derive(PartialEq)]
     struct Point {
         x: i32,
@@ -764,7 +764,7 @@ fn test_field_setter_transform() {
 }
 
 #[test]
-fn test_field_setter_transform_with_generics() {
+fn test_field_setter_transform_fn() {
     struct MBaseCase;
 
     struct MClosure;
@@ -796,7 +796,28 @@ fn test_field_setter_transform_with_generics() {
     struct Foo {
         #[builder(
             setter(
-                transform = #[generics(<'__a, __Marker>)] |value: impl IntoValue<'__a, String, __Marker>| value.into_value()
+                fn transform<'a, M>(value: impl IntoValue<'a, String, M>) -> String
+                where
+                    M: 'a,
+                 {
+                    value.into_value()
+                },
+            )
+        )]
+        s: String,
+    }
+
+    // Check where clause
+    #[derive(TypedBuilder)]
+    struct Bar {
+        #[builder(
+            setter(
+                fn transform<A>(value: A) -> String
+                where
+                    A: std::fmt::Display,
+                 {
+                    value.to_string()
+                },
             )
         )]
         s: String,
@@ -804,6 +825,8 @@ fn test_field_setter_transform_with_generics() {
 
     assert_eq!(Foo::builder().s("foo").build().s, "foo".to_owned());
     assert_eq!(Foo::builder().s(|| "foo".to_owned()).build().s, "foo".to_owned());
+
+    assert_eq!(Bar::builder().s(42).build().s, "42".to_owned());
 }
 
 #[test]