+++ typed-builder-macro/src/struct_info.rs
@@ -315,18 +315,22 @@ impl<'a> StructInfo<'a> {
             }
         });
 
-        let (method_generics, param_list, arg_expr) = if field.builder_attr.setter.strip_bool.is_some() {
-            (quote!(), quote!(), quote!(true))
+        let (method_generics, param_list, arg_expr, method_where_clause) = if field.builder_attr.setter.strip_bool.is_some() {
+            (quote!(), quote!(), quote!(true), quote!())
         } else if let Some(transform) = &field.builder_attr.setter.transform {
             let params = transform.params.iter().map(|(pat, ty)| quote!(#pat: #ty));
             let body = &transform.body;
             let method_generics = transform.generics.as_ref().map_or(quote!(), |g| g.to_token_stream());
-
-            (method_generics, quote!(#(#params),*), quote!({ #body }))
+            let method_where_clause = transform
+                .generics
+                .as_ref()
+                .and_then(|g| g.where_clause.as_ref())
+                .map_or(quote!(), |w| w.to_token_stream());
+            (method_generics, quote!(#(#params),*), quote!({ #body }), method_where_clause)
         } else if option_was_stripped {
-            (quote!(), quote!(#field_name: #arg_type), quote!(Some(#arg_expr)))
+            (quote!(), quote!(#field_name: #arg_type), quote!(Some(#arg_expr)), quote!())
         } else {
-            (quote!(), quote!(#field_name: #arg_type), arg_expr)
+            (quote!(), quote!(#field_name: #arg_type), arg_expr, quote!())
         };
 
         let repeated_fields_error_type_name = syn::Ident::new(
@@ -346,7 +350,9 @@ impl<'a> StructInfo<'a> {
                 #deprecated
                 #doc
                 #[allow(clippy::used_underscore_binding, clippy::no_effect_underscore_binding)]
-                pub fn #method_name #method_generics (self, #param_list) -> #builder_name <#target_generics> {
+                pub fn #method_name #method_generics (self, #param_list) -> #builder_name <#target_generics>
+                #method_where_clause
+                {
                     let #field_name = (#arg_expr,);
                     let ( #(#destructuring,)* ) = self.fields;
                     #builder_name {
@@ -364,7 +370,9 @@ impl<'a> StructInfo<'a> {
                 #deprecated
                 #doc
                 #[allow(clippy::used_underscore_binding, clippy::no_effect_underscore_binding)]
-                pub fn #method_name #method_generics (self, #param_list) -> #builder_name <#target_generics> {
+                pub fn #method_name #method_generics (self, #param_list) -> #builder_name <#target_generics>
+                #method_where_clause
+                {
                     let #field_name = (#arg_expr,);
                     let ( #(#destructuring,)* ) = self.fields;
                     #builder_name {
@@ -384,7 +392,9 @@ impl<'a> StructInfo<'a> {
                 #deprecated
                 #doc
                 #[allow(clippy::used_underscore_binding, clippy::no_effect_underscore_binding)]
-                pub fn #method_name #method_generics (self, #param_list) -> #builder_name <#target_generics> {
+                pub fn #method_name #method_generics (self, #param_list) -> #builder_name <#target_generics>
+                #method_where_clause
+                {
                     let #field_name = (#arg_expr,);
                     let ( #(#destructuring,)* ) = self.fields;
                     #builder_name {
@@ -407,7 +417,9 @@ impl<'a> StructInfo<'a> {
                     note = #repeated_fields_error_message
                 )]
                 #doc
-                pub fn #method_name #method_generics (self, _: #repeated_fields_error_type_name) -> #builder_name <#target_generics> {
+                pub fn #method_name #method_generics (self, _: #repeated_fields_error_type_name) -> #builder_name <#target_generics>
+                #method_where_clause
+                {
                     self
                 }
             }