+++ CHANGELOG.md
@@ -7,7 +7,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
-- `#[generics(<'a, A, B, ...>)]` can now be added as an attribute to the transform closure, e.g. `#[builder(setter(transform = #[generics(<A>)] |value: impl Foo<A>| expr))]`, to allow adding custom generics, bounds and lifetimes to the builder method.
+- New optional alternate `transform` syntax using a full fn, to allow support for custom lifetimes, generics and a where clause to custom builder method.
+
+Example:
+```rust
+#[derive(TypedBuilder)]
+struct Foo {
+    #[builder(
+        setter(
+            fn transform<'a, M>(value: impl IntoValue<'a, String, M>) -> String
+            where
+              M: std::fmt::Display
+            {
+                value.into_value()
+            },
+        )
+    )]
+    s: String,
+}
+```
 
 ## 0.21.2 - 2025-08-21
 ### Fixed